+++ packages/core/src/types/callback/index.ts
@@ -1,3 +1,2 @@
 export * from './action'
 export * from './contractVerifier'
-export * from './offChainAction'